requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
flasgger==0.9.7.1
PyYAML==6.0.2
python-dotenv==1.0.1
pymysql==1.1.1
cryptography==46.0.3
